업로드 재수정

+++ api/drive/api.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 )
 
-func Init( e *echo.Echo) {
+func Init(e *echo.Echo) {
 	g := e.Group("/api")
 
 	g.GET("/drives", GetDriveInfo)
@@ -19,8 +19,7 @@ func Init( e *echo.Echo) {
 
 }
 
-
-func GetDriveInfo (c echo.Context) error {
+func GetDriveInfo(c echo.Context) error {
 	cms := c.(*env.GoCms)
 
 	results, err := cms.Drives.Drives.List().Fields("kind, nextPageToken, drives").Do()
@@ -32,11 +31,10 @@ func GetDriveInfo (c echo.Context) error {
 	return c.JSON(200, results)
 }
 
-
 /*
   모든 파일 가져오기
   폴더, 파일이든 모드 파일로 취급
- */
+*/
 func GetFilesInDrive(c echo.Context) error {
 	cms := c.(*env.GoCms)
 
@@ -54,7 +52,6 @@ func GetFilesInDrive(c echo.Context) error {
 
 }
 
-
 func GetFileDetails(c echo.Context) error {
 	id := c.Param("id")
 	cms := c.(*env.GoCms)
@@ -64,7 +61,6 @@ func GetFileDetails(c echo.Context) error {
 		Fields("id, kind, mimeType").
 		Do()
 
-
 	if err != nil {
 		return c.JSON(500, err)
 	} else {
@@ -73,47 +69,43 @@ func GetFileDetails(c echo.Context) error {
 }
 
 type FileRequest struct {
-	Name     string `json:"name" validate:"required"`
-	MimeType string `json:"mimeType,omitempty" validate:"required"`
-	Parents []string `json:"parents,omitempty"`
-	Email string `json:"email,omitempty" validate:"email"`
+	Name     string   `json:"name" validate:"required"`
+	MimeType string   `json:"mimeType,omitempty" validate:"required"`
+	Parents  []string `json:"parents,omitempty"`
+	Email    string   `json:"email,omitempty" validate:"email"`
 }
 
+func CreateFile(c echo.Context) (err error) {
 
-
-
-func CreateFile( c echo.Context) (err error) {
-
-	var isMultipart = true
 	form, err := c.MultipartForm()
 
 	var fileRequest = new(FileRequest)
 
-	if  err != nil {
-		c.Logger().Error(err)
-		_ = c.Bind(&fileRequest)
-		isMultipart = false
-	}
-
-	getFirstValue := func( field []string) string {
-
-		if len(field) > 0 {
-			return strings.TrimSpace(field[0])
+	if err != nil { //form type이 multipart가 아니면 에러 발생
+		err := c.Bind(&fileRequest)
+		if err != nil {
+			return err
 		}
-		return ""
-	}
+	} else {
 
-	if isMultipart {
+		getFirstValue := func(field []string) string {
+
+			if len(field) > 0 {
+				return strings.TrimSpace(field[0])
+			}
+			return ""
+		}
 
 		fileRequest = &FileRequest{
-			Name: form.Value["name"][0],
+			Name:     form.Value["name"][0],
 			MimeType: getFirstValue(form.Value["mimeType"]),
-			Parents: form.Value["parents"],
-			Email: getFirstValue(form.Value["email"]),
+			Parents:  form.Value["parents"],
+			Email:    getFirstValue(form.Value["email"]),
 		}
 	}
 
-	err = c.Validate(fileRequest); if err != nil {
+	err = c.Validate(fileRequest)
+	if err != nil {
 		return c.JSON(400, err)
 	}
 
@@ -122,7 +114,6 @@ func CreateFile( c echo.Context) (err error) {
 	googleFile := &drive.File{
 		Name: fileRequest.Name,
 	}
-
 
 	if fileRequest.MimeType != "" {
 		googleFile.MimeType = fileRequest.MimeType
@@ -133,8 +124,9 @@ func CreateFile( c echo.Context) (err error) {
 	}
 
 	var apiResult *drive.File
+	var apiErr error
 
-	if isMultipart && form != nil {
+	if form != nil {
 
 		file := form.File["files"][0:1][0]
 
@@ -146,66 +138,57 @@ func CreateFile( c echo.Context) (err error) {
 
 		defer src.Close()
 
-		result, err := cms.Drives.Files.
-		Create(googleFile).
-		Media(src).
-		ProgressUpdater(func(current, total int64) {
-			fmt.Printf("-> %d, %d\r", current, total)
-
-		}).Do()
-
-		if err != nil {
-			return c.JSON(500, err)
-		}
-
-
-		apiResult = result
+		apiResult, apiErr = cms.Drives.Files.
+			Create(googleFile).
+			Media(src).
+			ProgressUpdater(func(current, total int64) {
+				fmt.Printf("UPLOAD PROGRESS -> %d, %d\r", current, total) //TODO 콜백처리
+			}).Do()
 
 	} else {
-
-		result, err := cms.Drives.Files.Create(googleFile).Do()
-
-		if err != nil {
-			return c.JSON(500, err)
-		}
-
-		apiResult = result
+		apiResult, apiErr = cms.Drives.Files.Create(googleFile).Do()
 	}
 
+	if apiErr != nil {
+		return c.JSON(500, apiErr)
+	}
 
 	if fileRequest.Email != "" {
-		err := NewPermission(cms, apiResult.Id, fileRequest.Email)
+		_, err := NewPermission(cms, apiResult.Id, fileRequest.Email, "owner")
 		if err != nil {
 			cms.Drives.Files.Delete(apiResult.Id)
 			cms.Logger().Debugf("giving a new permission - %s was failed", fileRequest.Email)
-			c.JSON(500, err)
+			return c.JSON(500, err)
 		}
 	}
 
 	return c.JSON(200, apiResult)
 
-
 }
 
-
-func NewPermission(cms *env.GoCms,  fileId string, email string) error {
-	_, err := cms.Drives.Permissions.Create(fileId, &drive.Permission{
+func NewPermission(cms *env.GoCms, fileId string, email string, role string) (*drive.Permission, error) {
+	result, err := cms.Drives.Permissions.Create(fileId, &drive.Permission{
 		EmailAddress: email,
-		Type: "user",
-		Role: "reader"}).
-		TransferOwnership(false).
+		DisplayName:  email,
+		Type:         "user",
+		Role:         role}).
+		TransferOwnership(func() bool {
+			if role == "owner" {
+				return true
+			}
+			return false
+		}()).
 		SupportsAllDrives(true).
 		SendNotificationEmail(true).Do()
 
 	if err != nil {
-		return err
+		return nil, err
 	}
 
-	return nil
+	return result, nil
 }
 
-
-func DeleteFile( c echo.Context) error {
+func DeleteFile(c echo.Context) error {
 	id := c.Param("id")
 	cms := c.(*env.GoCms)
 	err := cms.Drives.Files.Delete(id).Do()
@@ -215,65 +198,4 @@ func DeleteFile( c echo.Context) error {
 	} else {
 		return c.NoContent(200)
 	}
-}
-
-
-func ShareFile( c echo.Context) error  {
-	id := c.Param("id")
-	cms := c.(*env.GoCms)
-
-	result, err := cms.Drives.Permissions.Create(id, &drive.Permission{
-
-		EmailAddress: "clonekim@gmail.com",
-		//DisplayName: "clonekim@gmail.com",
-		Type: "user",
-		Role: "owner",
-		//AllowFileDiscovery: true,
-		//Domain: "hist.co.kr",
-		//AllowFileDiscovery: true,
-	}).TransferOwnership(true).SupportsAllDrives(true).SendNotificationEmail(true) .Do()
-
-	if err != nil {
-		return c.JSON(500, err)
-	} else {
-		return c.JSON(200, result)
-	}
-}
-
-
-
-func Upload( c echo.Context) error {
-
-	form, err := c.MultipartForm()
-
-	if  err != nil {
-		return err
-	}
-
-
-	file := form.File["files"][0:1][0]
-	src, err := file.Open()
-
-	if err != nil {
-		return err
-	}
-
-	cms := c.(*env.GoCms)
-	drivefile := &drive.File{Name: file.Filename}
-	result, err := cms.Drives.Files.
-		Create(drivefile).
-		Media(src).
-		ProgressUpdater(func(current, total int64) {
-		  fmt.Printf("-> %d, %d\r", current, total)
-
-	}).Do()
-
-	defer src.Close()
-
-	if err != nil {
-		return c.JSON(500, err)
-	}
-
-	return c.JSON(200, result)
-
 }
